pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = "Contact Management System"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "fastapi-mail",
    "cloudinary",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
    "."
]

